Fixed compile bugs.
Former-commit-id: 6c07a51ade2bc968ef2a179872c8254304784fc3

+++ modules/grabberServer/cxGrabber.h
@@ -14,7 +14,7 @@ namespace cx
 class Frame
 {
 public:
-  float mTimestamp; ///< Timestamp in milliseconds
+  double mTimestamp; ///< Timestamp in milliseconds
   int mWidth; ///< Width in pixels
   int mHeight; ///< Height in pixels
   int mPixelFormat; ///< Pixel format in OSType (FourCC)
